refactor: Api Controller 클래스 구조 리팩토링
클래스 책임에 맞게 클래스 구조를 리팩토링
